+++ package_handler.py
@@ -700,7 +700,7 @@ class PackageHandler:
         output = result.stdout.splitlines()
 
         for line in output:
-            if line.startswith("Architecture"):  # TODO: Currently language-dependent
+            if line.startswith(self.config.config.get("architecture-wording")):
                 package_architecture = line.split(":")[1].strip()
                 self.logger.debug(f"Package architecture: {package_architecture}")
                 break