+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **Changelog** Added error message when no major origin package changelog was found
 - **Changelog** Added package list in changelog output
 - **Menu** Added a package changelog selection menu to check either single, multiple or all packages
+- **Config** Added the option to change the package architecture dependent word in the config instead of in the code
 
 ### Bug fixes